refactor: split ast analyzer models
